requirements.txt
line-bot-sdk==1.8.0
flask
selenium==3.141.0
pytz==2019.1
gunicorn
psycopg2-binary==2.8.3
pyowm
mistune==0.8.4
numpy
pandas==0.24.0
requests==2.22.0
lxml
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
oauth2client
openai